Fixed compiler warnings.

+++ include/mrf/downsample.hpp
@@ -32,7 +32,7 @@ template <typename T>
 const typename pcl::PointCloud<T>::Ptr downsampleRandom(
     const typename pcl::PointCloud<T>::ConstPtr& in, const double& random_rate) {
     const typename pcl::PointCloud<T>::Ptr out{new typename pcl::PointCloud<T>};
-    const size_t num_points{random_rate * in->size()};
+    const size_t num_points{(size_t)(random_rate * in->size())};
     out->reserve(num_points);
 
     std::srand(std::time(0));
